requirements.txt
annotated-types==0.6.0
asgiref==3.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==5.0.4
django-choices-field==2.3.0
django-cors-headers==4.3.1
django-environ==0.9.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.15.1
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.12.0
googleapis-common-protos==1.65.0
graphql-core==3.2.3
gunicorn==23.0.0
grpcio==1.62.1
idna==3.6
importlib-metadata==6.11.0
numpy==1.26.4
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-django==0.44b0
opentelemetry-instrumentation-wsgi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
packaging==24.0
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
pyecharts==2.0.3
pyecharts-snapshot @ git+https://github.com/Deepthi-Chand/pyecharts-snapshot.git@8d6cadd055db6c919a1447064185d00d1b30ce01
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
strawberry-graphql==0.225.1
strawberry-graphql-django==0.37.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.6
wrapt==1.16.0
zipp==3.18.1
snapshot-selenium==0.0.2
selenium==4.25.0

# Security and Environment Management
python-decouple==3.8
django-debug-toolbar==4.3.0
django-ratelimit==4.1.0

# Caching and Performance
django-redis==5.4.0
django-cacheops==7.0.2

# Code Quality and Type Checking
mypy==1.15.0
black==24.2.0
isort==5.13.2
pre-commit==3.6.2
flake8==7.0.0
pandas-stubs==2.2.3.241126

# Logging
structlog==24.1.0

# Development and Debugging
ipython==8.22.1

# API Documentation
drf-yasg==1.21.7
drf-yasg[validation]==1.21.7
djangorestframework-simplejwt==5.3.1

# API Versioning and Throttling
django-filter==24.1

django-stubs==5.1.3
djangorestframework-stubs==3.15.3
